feat(cli): add AgentHarness trait, HostHarness adapter, and worker environment validation (VLC-9) (#8)
- Implement AgentHarness trait with HarnessResult and ResourceLimits
- Add domain error types via thiserror: InstallError, VmBootError, VmTimeout, OpenCodeCrash, OutputParseError
- Implement HostHarness: runs opencode on host, parses PR URL and token usage from stdout
- Add worker environment validation (KVM, binaries, rootfs) with Severity levels
- Add 'vulcanum worker setup' subcommand
- Wire HostHarness into daemon handle_job with real result submission
- Daemon startup now validates environment before entering poll loop
- Add unit tests for error types, PR URL parser, and token usage parser

+++ cli/src/client.rs
@@ -55,8 +55,8 @@ pub struct JobResponse {
 pub struct SubmitResultRequest {
     pub pr_url: String,
     pub exit_code: i32,
-    pub tokens_used: i32,
-    pub duration_ms: i32,
+    pub tokens_used: u64,
+    pub duration_ms: u64,
 }
 
 pub struct ApiClient {
